refac: refactor consumer.py file name to consumers.py
